Refactor to smaller components
